--- a/be/src/column/json_column.h
+++ b/be/src/column/json_column.h
@@ -113,9 +113,15 @@ public:
 
     Columns& get_flat_fields() { return _flat_columns; };
 
+    const Columns& get_flat_fields() const { return _flat_columns; };
+
     ColumnPtr& get_flat_field(int index);
 
     const ColumnPtr& get_flat_field(int index) const;
+
+    ColumnPtr& get_remain();
+
+    const ColumnPtr& get_remain() const;
 
     const std::vector<std::string>& flat_column_paths() const { return _flat_column_paths; }
 
@@ -123,17 +129,18 @@ public:
 
     bool has_flat_column(const std::string& path) const;
 
-    void init_flat_columns(const std::vector<std::string>& paths);
+    bool has_remain() const { return _flat_columns.size() == (_flat_column_paths.size() + 1); }
 
-    void init_flat_columns(const std::vector<std::string>& paths, const std::vector<LogicalType>& types);
+    void set_flat_columns(const std::vector<std::string>& paths, const std::vector<LogicalType>& types,
+                          const Columns& flat_columns);
 
     std::string debug_flat_paths() const;
 
 private:
-    // flat-columns
+    // flat-columns[sub_columns, remain_column]
     Columns _flat_columns;
 
-    // flat-column paths
+    // flat-column paths, doesn't contains remain column
     std::vector<std::string> _flat_column_paths;
     std::vector<LogicalType> _flat_column_types;
     std::unordered_map<std::string, int> _path_to_index;