--- a/be/src/storage/olap_common.h
+++ b/be/src/storage/olap_common.h
@@ -302,7 +302,11 @@ struct OlapReaderStatistics {
     // ------ for json type, to count flat column ------
     // key: json absolute path, value: count
     int64_t json_flatten_ns = 0;
+    int64_t json_cast_ns = 0;
+    int64_t json_merge_ns = 0;
+    int64_t json_init_ns = 0;
     std::unordered_map<std::string, int64_t> flat_json_hits;
+    std::unordered_map<std::string, int64_t> merge_json_hits;
     std::unordered_map<std::string, int64_t> dynamic_json_hits;
 };