## Why I'm doing:
Before this PR, starrocks only support one-layer flat json, and will save origin data.
This PR is support multi-layer flat json, and doesn't save origin data.

## What I'm doing:
Will flat all leaf column in json column, the json meta will be:
![image](https://github.com/StarRocks/starrocks/assets/5609319/23d0c035-ceb5-43a3-9002-1349cac79d33)

the core file:
* `json_flattener.h`: to handle json flatten, json merge process...
* `json_column_iterator.h`: to handle read flat json, read whole json...
* `json_column_compactor.h`: to handle compaction flat json proces...
* `json_column_writer.h`: the logical is same as before

### Test result:
this pr for optimize load&data size
|  | Scalar Data | Json | Flat Json(Before PR) | FlatJson（This PR) |
| -- | -- | -- | -- | -- |
| time(s) | 373.069s | 687.961s | 1617.537s | 991.582s |
| Data Size | 58.828 GB | 152.616GB | 245.947 GB | 58.571 GB |
| compaction time | 4min30s | 35min20s | 46min51s | 23 min 30s |

because compaction task depend on `HeapMerge` or `MaskMerge`, we can't read flat json direct (may schema different), will merge all subfield to json to do compaction

### Write
* remove origin json data, only save flat json and remain data

flat json storage data&order:
* null data
* flat subfield data
* remain data

### Compaction (large different)
support multi mode compaction:

* read json and compaction to json (now)
* read json and compaction to flat json (now)
  * flatten json
* read flat json and compaction to flat json (can't use)
  * match type and path
  * cast flat json type to target
  * merge flat json to remain
  * extract flat json from remain 
* read flat json and compaction to json (can't use)
  * merge flat json to json

### Read (large different)
update read logical:

* support read from flat json
  * when read flat json
    * cast from remain json
    * extract from remain json or flat json
    * merge flat json and remain json, maybe need cast
    * when can't hit any flat json, we need return empty column 
  * when read whole json
    * merge all flat json 
* support read from json
  * when read flat json
    * dynamic flattern json 
  * when read json, return directly


## What type of PR is this:

core changes:
* refactor `JsonFlattern` class:
  * refactor `JsonFlattern`, split path deriver into `JsonPathDeriver`, split json flatten to `JsonFlattener`
  * add `JsonMerger`, for merge any flat json and remain to a json binary(vpack json)
    * support different mode:  
      * reset new root path: for read process, like flat json is `$.a.b`, `$.a.c`, but we read `$.a`, we need merge flat json and rebase the root to `$.a`
      * set exclude path: for compaction process, some json will be extract from remain, but some will be merge into remain, we need remove extract path from remain
  * support `HyperJsonTransfor`: for read&compaction, to split the merge task and flat task base on input json
* refactor `JsonColumnWriter`:
  * split write process to `JsonColumnWriter`, split compaction process to `JsonColumnCompactor`
  * support `FlatJsonColumnCompactor`  to handle compaction flat json to flat json
  * support `JsonColumnCompactor` to handle compaction flat json to json
* refactor `JsonColumnIterator`:
  * refactor `JsonFlatIterator` to handle read flat json
  * support `JsonMergeIterator` to handle read whole json 
  * `JsonDynamicIterator`: keep same
* update `SegmentWriteOptions` , add `is_compaction` to mark is compaction process
  * update OlapTable procees
  * update LakeTable process 
* FE update Json subfiled deriver limit

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5