@stdpain 2024-07-31
we should make json_each works on flat json
```
    for (int i = 0; i < num_input_rows; i++) {
        const JsonValue* json = json_column->get_object(i);
        DCHECK(!!json);
        vpack::Slice json_slice = json->to_vslice();
        if (json_slice.isObject()) {
            for (auto [key, value] : vpack::ObjectIterator(json_slice)) {
                std::string_view key_str = key.stringView();
                key_column_ptr->append(Slice(key_str.data(), key_str.size()));
                value_column_ptr->append(JsonValue(value));
                offset++;
            }
        } else if (json_slice.isArray()) {
            int arr_idx = 0;
            for (auto value : vpack::ArrayIterator(json_slice)) {
                std::string key = std::to_string(arr_idx);
                key_column_ptr->append(Slice(key));
                value_column_ptr->append(JsonValue(value));
                offset++;
                arr_idx++;
            }
        }
        // process flat values here 

        offset_column->append(offset);
    }
```