@Seaven 2024-07-31
> for (auto [key, value] : vpack::ObjectIterator(json_slice)) {

don't worry, flat json never meet `json_each`