@Seaven 2024-08-01
> When will this feature be available

wait this PR merge~